requirements.txt
kafka-python
requests
python-dotenv
pydub
huggingface_hub
langdetect
aiohttp
aiokafka
boto3
supabase
